pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.9.2",
    "docling>=2.64.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
